mockup service to manipulate user uploaded images
