+++ .github/workflows/mockup-staging.yaml
@@ -1,4 +1,4 @@
-name: fetch-service staging
+name: mockup-service staging
 
 on:
   workflow_dispatch:
@@ -6,7 +6,7 @@ on:
     branches:
       - 'staging'
     paths:
-      - services/fetch-service/**
+      - services/mockup-service/**
 
 jobs:
   build:
@@ -23,13 +23,13 @@ jobs:
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: Fetch Service
+      - name: mockup Service
         uses: docker/build-push-action@v2
         with:
-          context: services/fetch-service
+          context: services/mockup-service
           platforms: linux/amd64,linux/arm64
           push: true
-          tags: bitprj/tinyhat-fetch-service:beta
+          tags: bitprj/tinyhat-mockup-service:beta
           
   deploy:
     needs: build
@@ -69,7 +69,7 @@ jobs:
         uses: rtCamp/action-slack-notify@master                 
         env:                                                    
           SLACK_CHANNEL: 'k8s-work'                  
-          SLACK_MESSAGE: 'Just deployed fetch-service to staging!'
+          SLACK_MESSAGE: 'Just deployed mockup-service to staging!'
           SLACK_TITLE: 'Deploy'                         
           SLACK_USERNAME: 'Daniel'                           
           SLACK_ICON: "https://ca.slack-edge.com/TJ2FB5CE9-UHW7UUSDP-6f70027657b3-512"
