--- kube-local/README.md
@@ -1,1 +1,0 @@
-Kubernetes Manifests for deploying locally on `minikube`.
