--- kube-local/add.yaml
@@ -1,37 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: add-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: add-service
-  template:
-    metadata:
-      labels:
-        app: add-service
-    spec:
-      containers:
-        - name: add-service
-          image: bitprj/mini-add-service
-          ports:
-            - containerPort: 31337
-          env:
-            - name: UPLOAD_ENDPOINT
-              value: upload-service:80
-            - name: ENDPOINT
-              value: "$ENDPOINT"
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: add-service
-spec:
-  selector:
-    app: add-service
-  ports:
-    - port: 80
-      targetPort: 31337
-  type: ClusterIP
