--- kube-local/admin.yaml
@@ -1,37 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: admin-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: admin-service
-  template:
-    metadata:
-      labels:
-        app: admin-service
-    spec:
-      containers:
-        - name: admin-service
-          image: bitprj/mini-admin-service
-          ports:
-            - containerPort: 4040
-          env:
-            - name: HOST
-              value: mysql
-            - name: PASSWORD
-              value: password
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: admin-service
-spec:
-  selector:
-    app: admin-service
-  ports:
-    - port: 80
-      targetPort: 4040
-  type: ClusterIP
