--- kube-local/fetch.yaml
@@ -1,39 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: fetch-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: fetch-service
-  template:
-    metadata:
-      labels:
-        app: fetch-service
-    spec:
-      containers:
-        - name: fetch-service
-          image: bitprj/mini-fetch-service
-          ports:
-            - containerPort: 1337
-          env:
-            - name: HOST
-              value: mysql
-            - name: PASSWORD
-              value: password
-            - name: MANIPULATE_ENDPOINT
-              value: manipulation-service:80
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: fetch-service
-spec:
-  selector:
-    app: fetch-service
-  ports:
-    - port: 80
-      targetPort: 1337
-  type: ClusterIP
