--- kube-local/frontend.yaml
@@ -1,32 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: frontend-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: frontend-service
-  template:
-    metadata:
-      labels:
-        app: frontend-service
-    spec:
-      containers:
-        - name: frontend-service
-          image: bitprj/mini-frontend-service
-          ports:
-            - containerPort: 3000
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: frontend-service
-spec:
-  selector:
-    app: frontend-service
-  ports:
-    - port: 80
-      targetPort: 3000
-  type: LoadBalancer
