--- kube-local/gateway.yaml
@@ -1,39 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: gateway-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: gateway-service
-  template:
-    metadata:
-      labels:
-        app: gateway-service
-    spec:
-      containers:
-        - name: gateway-service
-          image: bitprj/mini-gateway-service
-          ports:
-            - containerPort: 4444
-          env:
-            - name: FETCH_ENDPOINT
-              value: fetch-service:80
-            - name: ADD_ENDPOINT
-              value: add-service:80
-            - name: MODERATE_ENDPOINT
-              value: moderate-service:80
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: gateway-service
-spec:
-  selector:
-    app: gateway-service
-  ports:
-    - port: 80
-      targetPort: 4444
-  type: LoadBalancer
