--- kube-local/moderate.yaml
@@ -1,37 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: moderate-service
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: moderate-service
-  template:
-    metadata:
-      labels:
-        app: moderate-service
-    spec:
-      containers:
-        - name: moderate-service
-          image: bitprj/mini-moderate-service
-          ports:
-            - containerPort: 5000
-          env:
-            - name: HOST
-              value: mysql
-            - name: PASSWORD
-              value: password
-          imagePullPolicy: Always
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: moderate-service
-spec:
-  selector:
-    app: moderate-service
-  ports:
-    - port: 80
-      targetPort: 5000
-  type: ClusterIP
